V3.3.1

+++ src/column-pack/keys/big-int-key-column.ts
@@ -2,7 +2,7 @@ import { ColumnOptions, ColumnType } from '../../column';
 
 export const BigIntKeyColumn: ColumnOptions = {
 	name: 'id',
-	type: ColumnType.INT,
+	type: ColumnType.BIGINT,
 	primaryKey: true,
 	autoIncrement: true,
 };
